+++ CHANGELOG.md
@@ -9,6 +9,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 - Described and implemented protocol for dao recommendations
+- Export DAO popularity index
 
 ## [0.1.12] - 2024-04-11