+++ CHANGELOG.md
@@ -7,6 +7,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [0.1.13] - 2024-04-19
+
 ### Added
 - Described and implemented protocol for dao recommendations
 - Export DAO popularity index